Movement Components
Components which will save movement-related info.

Moving is applied to any moving entity and contains a vector of a direction it is moving in. This vector is not normalized, it implicitly contains current movement speed. It's (0,0,0) when idle.

CharacterSpeeds is a component to be used for any moving characters in the game. It controls their speeds in certain directions.
- run_speed - speed when moving left-right
- jump_speed - added speed in the upwards direction when jump command is pressed, then it will be slowly brought down by gravity
- dash_speed - speed you reach when dashing left-right, vertical axis is locked while you do this